+++ src/components/shared/ColorCell.tsx
@@ -1,6 +1,6 @@
 import React, {ReactNode} from 'react'
 import styled, {keyframes} from "styled-components"
-import grid from "../../assets/no-color-grid-small.png"
+import grid from "../../../public/images/no-color-grid-small.png"
 
 interface ColorCellProps {
   outline?: boolean
